move TestLocationOwnership to own file
