+++ corehq/apps/fixtures/tests/__init__.py
@@ -1,2 +1,3 @@
 from .tests import *
 from .test_fixture_data import *
+from .test_location_ownership import *
